+++ examples/locales/en-US/resources.js
@@ -377,6 +377,7 @@ window.resources = {
         "title_findMTSPPathsServiceLogistics": "Multi-traveler Analysis - Logistics Distribution ",
         "title_findTSPPathsService": "Multi-traveler Analysis - Quality Inspection",
         "title_graphicLayer":"High efficiency point layer - points of 180K NY taxis",
+        "title_mb_graphicLayer":"High efficiency point layer",
         "title_graphicLayerClovers":"High efficiency point layer - clover",
         "title_graphicLayerTaxi":"High efficiency point layer - picture drawing of 180K NY taxis",
         "title_heatMap": "Heat map",