+++ examples/locales/zh-CN/resources.js
@@ -377,6 +377,7 @@ window.resources = {
         "title_findMTSPPathsServiceLogistics":"网络分析服务-多旅行商分析(物流配送)",
         "title_findTSPPathsService":"网络分析服务-多旅行商分析(质检巡查)",
         "title_graphicLayer":"高效率点图层-18万纽约出租车点",
+        "title_mb_graphicLayer":"高效率点图层",
         "title_graphicLayerClovers":"高效率点图层-三叶草",
         "title_graphicLayerTaxi":"高效率点图层-18万纽约出租车图片绘制",
         "title_heatMap":"热力图",