+++ src/mapboxgl/core/Util.js
@@ -96,6 +96,32 @@ export class Util {
         return geometryParam;
     }
 
+    /**
+     * @function mapboxgl.supermap.Util.extend
+     * @description 对象拷贝赋值
+     * @param dest - 目标对象
+     * @param arguments - 待拷贝的对象
+     * @return 赋值后的目标对象
+     */
+    static extend(dest) {
+        for (var index = 0; index < Object.getOwnPropertyNames(arguments).length; index++) {
+            var arg = Object.getOwnPropertyNames(arguments)[index];
+            if (arg == "caller" || arg == "callee" || arg == "length" || arg == "arguments") {
+                continue;
+            }
+            var obj = arguments[arg];
+            if (obj) {
+                for (var j = 0; j < Object.getOwnPropertyNames(obj).length; j++) {
+                    var key = Object.getOwnPropertyNames(obj)[j];
+                    if (arg == "caller" || arg == "callee" || arg == "length" || arg == "arguments") {
+                        continue;
+                    }
+                    dest[key] = obj[key];
+                }
+            }
+        }
+        return dest;
+    }
 }
 
 mapboxgl.supermap.Util = Util;