+++ src/mapboxgl/overlay/index.js
@@ -1,3 +1,4 @@
+
 export {Graph} from './GraphThemeLayer';
 export {Label} from './LabelThemeLayer';
 export {MapvLayer} from './MapvLayer';
@@ -6,8 +7,10 @@ export {Range} from './RangeThemeLayer';
 export {RankSymbol} from './RankSymbolThemeLayer';
 export {UniqueTheme3DLayer} from './UniqueTheme3DLayer';
 export {Unique} from './UniqueThemeLayer';
+export {GraphicLayer} from "./GraphicLayer";
 export {ThreeLayer} from './ThreeLayer';
 export {HeatMapLayer} from './HeatMapLayer';
+export * from './graphic';
 export * from './mapv';
 export * from './theme';
 export * from './threejs';